fixup! Add script for issue 3556
