+++ test_sets/Defects/Defects_release_7_1.txt
@@ -25,5 +25,5 @@
 ./test_scripts/Defects/7_1/3469_UI_SetGlobalProperties_during_resumption.lua
 ./test_scripts/Defects/7_1/3409_1_Single_transport.lua
 ./test_scripts/Defects/7_1/3409_2_Multiple_transports.lua
-./test_scripts/Defects/7_1/3556_1_GetListOfPermission.lua
-./test_scripts/Defects/7_1/3556_2_GetListOfPermission.lua
+./test_scripts/Defects/7_1/3556_1_GetListOfPermission_no_specific_permission.lua
+./test_scripts/Defects/7_1/3556_2_GetListOfPermission_contains_groups_with_user_consent_prompt.lua
